Refactor divider
