+++ src/js/addons/UI.js
@@ -1,6 +1,6 @@
 import './ui/file-drop';
+import { initDivider } from './ui/divider';
 import Menu from './ui/Menu';
-import Divider from './ui/Divider';
 import Tooltip from './ui/Tooltip';
 import MapToolbar from './ui/MapToolbar';
 
@@ -10,8 +10,9 @@ export default class UI {
         Tooltip.init();
         MapToolbar.init();
 
+        initDivider();
+
         // TODO: less inheritance
         this.menu = new Menu();
-        new Divider();
     }
 }
